requirements.txt
numpy
Pillow
pytesseract
pytz
PyYAML
icalendar
requests